requirements.txt
amqp==5.2.0
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
async-timeout==4.0.3
Babel==2.14.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.2
django-simple-history==3.5.0
django-appconf==1.0.6
django-cryptography==1.1
djangorestframework==3.14.0
docxcompose==1.4.0
docxtpl==0.16.8
Jinja2==3.1.3
kombu==5.3.5
lxml==5.1.0
MarkupSafe==2.1.5
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
redis==5.0.2
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
factory_boy==3.3.0

